Net: Initialize link state at boot
I noticed a problem where the SMC95xx was reporting the wrong link
state via netlink on boot.  With no cable attached (and thus no link),
it would still report that there was a link present.  I believe this is
because of how netdevice no carrier is defined (it's a bit in the state
member called __LINK_STATE_NOCARRIER and is initially set to 0 by a
memset/kzalloc of the struct).  Nothing sets this flag currently until
a link_reset() is called (on plug) so the initial state is left wrong.

I added a call to mii_check_media() (what link_reset() does to set
carrier state) in the bind() routine to initialize the link state.
Tests shows this corrects the problem.

Change-Id: I4ffb350a9a75ef4385469ab2fde61502dcf80107
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>