(PARTIALLY) usb: ehci: Configure PHY-reset gpio_171
For the ehci port resume-error work-around, we need
to reset the USB 3320c Phy using a GPIO going into
the RESETB pin of the Phy.

With a Hardware modification on the tablet board of
having GPIO 171 feed into the RESETB pin of the ehci
phy, this patch configures GPIO_171 and uses it
to do the Phy reset along with the soft-reset of the
link(ehci) to recover from the resume-error.

Change-Id: Ie0eb5c0c6ed34fefd9c64a0f72ab89a1398f6bae
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>